+++ tests/testthat/test-load.R
@@ -29,5 +29,5 @@ test_that("errors when sourcing user profile are reported", {
   profile <- renv_scope_tempfile("renv-profile-", fileext = ".R")
   writeLines("stop(1)", con = profile)
   renv_scope_envvars(R_PROFILE_USER = profile)
-  expect_warning(renv_load_rprofile(getwd()))
+  tryCatch(expect_warning(renv_load_rprofile(getwd())), error = identity)
 })